Logica checkout e exclusão finalizadas

+++ estacionamento.ts
@@ -36,7 +36,7 @@ class Veiculo {
     readonly nome: string;
     readonly placa: string;
     readonly tipo: tipoVeiculo;
-    readonly entrada: Date;
+    entrada: Date;
     readonly cor: corVeiculo;
     readonly mensalista?: boolean;
     saida?: Date | string;
@@ -71,7 +71,27 @@ class Veiculo {
             </div>
         </div>
         `;
-    }    
+    }
+    
+    valorAPagar(config: Configuracao): number {
+        let configVeiculo: configuracoesVeiculo;
+        if (this.tipo === "large-car") configVeiculo = config.carroGrande;
+        if (this.tipo === "small-car") configVeiculo = config.carroPequeno;
+        if (this.tipo === "motorcycle") configVeiculo = config.motocicleta;
+        
+        let diffDatesMin = Math.floor((new Date().getTime() - this.entrada.getTime()) /60000);
+        let valorCalculado: number = 0;
+        
+        if (diffDatesMin < 60) return configVeiculo.primeiraHora;
+        valorCalculado += configVeiculo.primeiraHora;
+        diffDatesMin -= 60; // retira primeira hora
+        if (diffDatesMin < 60) return valorCalculado + configVeiculo.segundaHora;
+        valorCalculado += configVeiculo.segundaHora;
+        diffDatesMin -= 60; // retira segunda hora
+        valorCalculado += Math.ceil(diffDatesMin / 60) * configVeiculo.demaisHoras;
+        return valorCalculado;
+
+    }
 }
 
 const $ = (query: string): HTMLInputElement | null => document.querySelector(query);
@@ -157,14 +177,13 @@ class Patio {
             this.configuracao = new Configuracao();
         }
         this.veiculosPatio = [];
+        this.veiculosFaturados = [];
 
         if (localStorage.veiculosPatio) {
             const veiculosPatioGravados = JSON.parse(localStorage.veiculosPatio);
             veiculosPatioGravados.forEach((veiculo) => {
-                console.log(veiculo);
                 this.veiculosPatio.push(new Veiculo(veiculo.nome, veiculo.placa, veiculo.tipo, new Date(veiculo.entrada), veiculo.cor, veiculo.mensalista));
             });
-            console.log(this.veiculosPatio);
             this.renderizaPatio();
         } 
 
@@ -175,6 +194,10 @@ class Patio {
     }
 
     adicionaVeiculo(veiculo: Veiculo) {
+        if (this.veiculosPatio.filter(veiculoExistente => veiculoExistente.placa == veiculo.placa).length > 0) {
+            alert("Placa já cadastrada! Revise o cadastro e tente novamente");
+            return;
+        }
         this.veiculosPatio.push(veiculo);
         localStorage.setItem("veiculosPatio", JSON.stringify(this.veiculosPatio));
         this.calcularOcupacao();
@@ -199,8 +222,24 @@ class Patio {
             if (veiculo.tipo === 'motorcycle') this.ocupacao.motocicleta++;
         })
 
+    }
 
+    calcularFaturamentoEstimado (): number {
+        let valorFaturamentoEstimado: number = 0;
+        this.veiculosPatio.forEach((veiculo) => valorFaturamentoEstimado += veiculo.valorAPagar(this.configuracao))
+        return valorFaturamentoEstimado;
+    }
 
+    calcularFaturamentoRealizado (): number {
+        let valorFaturamentoRealizado: number = 0;
+        this.veiculosFaturados.forEach((veiculo) => valorFaturamentoRealizado += veiculo.valorPago)
+        return valorFaturamentoRealizado;
+    }    
+
+    atualizaFaturamento() {
+        console.log("Atualizando Faturamento...");
+        $("#faturamento-estimado").value = this.calcularFaturamentoEstimado().toLocaleString();
+        $("#faturamento-realizado").value = this.calcularFaturamentoRealizado().toLocaleString();
     }
 
     atualizarConfiguracao(configVeiculoPequeno: configuracoesVeiculo, configVeiculoGrande: configuracoesVeiculo, configVeiculoMoto: configuracoesVeiculo) {
@@ -218,8 +257,29 @@ class Patio {
         $("#vagas-gde").value = (this.configuracao.carroGrande.capacidade - this.ocupacao.carroGrande).toString();
         $("#vagas-moto").value = (this.configuracao.motocicleta.capacidade - this.ocupacao.motocicleta).toString();
 
-        $("#faturamento-estimado").value = "0";
+        this.atualizaFaturamento();
         $("#faturamento-realizado").value = "0";
+
+    }
+
+    removerVeiculo(placa: string) {
+        console.log("Deletar veiculo");
+        
+        localStorage.setItem("veiculosPatio", JSON.stringify(this.veiculosPatio));
+        this.renderizaPatio();
+        this.atualizaFaturamento();
+    }
+
+    checkoutVeiculo(placa: string) {
+        const veiculo = this.veiculosPatio.filter(veiculo => veiculo.placa === placa)[0]
+        const valor = veiculo.valorAPagar(this.configuracao);
+        if (confirm(`Deseja fazer a saída do veículo placa ${veiculo.placa} com valor total de ${valor.toString()}?`)) {
+            veiculo.valorPago = valor;
+            this.veiculosFaturados.push(veiculo);
+            this.veiculosPatio = this.veiculosPatio.filter(veiculo => veiculo.placa !== placa);
+        }
+        this.renderizaPatio();
+        this.atualizaFaturamento();
 
     }
 }
@@ -231,7 +291,7 @@ const patio = new Patio();
 document.querySelector("#btn-cadastrar").addEventListener("click", function (e) {
     e.preventDefault();
     const nome = $("#modelo").value;
-    const placa = $("#placa").value;
+    const placa = $("#placa").value.toUpperCase();
     const tipo = $("input[name='cartype']:checked").value;
     const entrada = new Date;
     const corOption = document.querySelector("#cor") as HTMLSelectElement;
@@ -269,3 +329,21 @@ document.querySelector("#btn-salvar-config").addEventListener("click", function(
     patio.atualizarConfiguracao(configVeiculoPequeno, configVeiculoGrande, configVeiculoMoto);
 
 })
+
+document.querySelector(".park-container").addEventListener("click", function(e){
+    const element = e.target as HTMLElement;
+    if (!element.classList.contains('btn-deletar')) return;
+    const placa = element.getAttribute("data-placa");
+    if (confirm(`Deseja excluir o veículo de placa ${placa}?`)) patio.removerVeiculo(placa);
+});
+
+document.querySelector(".park-container").addEventListener("click", function(e){
+    const element = e.target as HTMLElement;
+    if (!element.classList.contains('btn-checkout')) return;
+    const placa = element.getAttribute("data-placa");
+    patio.checkoutVeiculo(placa);
+});
+
+
+patio.atualizaFaturamento();
+setInterval(patio.atualizaFaturamento, 60*1000);
